odhcp6c: accept EUI64 and random for interface ID
ip6ifaceid is now properly passed to odhcp6c in OpenWrt scripts, which is
broken if we don't support 'eui64' and 'random' options.
Instead of printing help and exiting when an invalid fixed address is
configured, let's just log an error and pretend that no interface ID was
passed to odhcp6c.

Closes https://github.com/openwrt/openwrt/issues/20976

Signed-off-by: Álvaro Fernández Rojas <noltari@gmail.com>